fix: cast missingDays sum to int4, avoid string concat from unregistered int8 parser (CM-1463)

+++ services/libs/data-access-layer/src/repositoryStarSnapshots/index.ts
@@ -116,8 +116,8 @@ export async function findStarSnapshotGapDaysForRepos(
     `
       select
           "repositoryId",
-          ((now() at time zone 'UTC')::date - min(("capturedAt" at time zone 'UTC')::date) + 1)
-            - count(distinct ("capturedAt" at time zone 'UTC')::date) as "missingDays"
+          (((now() at time zone 'UTC')::date - min(("capturedAt" at time zone 'UTC')::date) + 1)
+            - count(distinct ("capturedAt" at time zone 'UTC')::date))::int as "missingDays"
       from "repositoryStarSnapshots"
       where "repositoryId" in ($(repositoryIds:csv))
       group by "repositoryId"
